jlink_spi: Increase delay on power on feature from 10 to 100 millisec
More delay time is needed to stabilize the LDO and the decoupling
capacitors when power on feature is enabled. Specifically 100 ms
was tested on HT7333 and it's sufficient.

I used the power supply from the J-Link device pin 19.
this power is 5V, I put a 3.3V LDO (HT7333) after it, and this LDO
(and decoupling capacitors) need more time to fix output voltage.

Change-Id: Ic2dd94e99ac6ffa17a009b8488ce027698ae2c28
Signed-off-by: Attila Veghelyi <aveghelyi@dension.com>
Tested-by: Attila Veghelyi <aveghelyi@dension.com>
Signed-off-by: Anastasia Klimchuk <aklm@flashrom.org>
Reviewed-on: https://review.coreboot.org/c/flashrom/+/86085
Reviewed-by: Anastasia Klimchuk <aklm@chromium.org>
Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>