fmap: Skip unreadable regions when attempting to locate FMAP
When using the on-chip FMAP to restrict flashrom operations to one or
more regions (--fmap), flashrom must first locate the FMAP. This
requires flashrom to read from multiple addresses, some of which
may be located in regions which are not readable (such as the
Intel ME region). In order to avoid a substantial amount of output from
read_flash() when trying to locate the FMAP in these regions, set
`the skip_unreadable_regions` flag before performing any reads to
locate the FMAP, and restore the original flag value when finished.

This resolves https://ticket.coreboot.org/issues/587

TEST=build flashrom, use cli to update COREBOOT and EC FMAP regions on
an Intel Alderlake-N platform board (starlabs/starlite_adl).

Change-Id: Ie78b977b4e6a5db02a25e69603f697834043ea99
Signed-off-by: Matt DeVillier <matt.devillier@gmail.com>
Reviewed-on: https://review.coreboot.org/c/flashrom/+/87748
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Anastasia Klimchuk <aklm@chromium.org>