doc: Add page to publish license info on the website
COPYING file existed in the tree for a very long time, but now
we also have a website which is generated from the docs in the tree.

This patch adds a page to display content of the COPYING file on
the website. The file stays in the tree, it just got slight
formatting added to convert it to rst format. No changes to the
text, just formatting (headings, paragraphs etc).

Change-Id: Ieb05d61ccc1b66e9ce91b7dfa18672a4da7db52c
Signed-off-by: Anastasia Klimchuk <aklm@flashrom.org>
Reviewed-on: https://review.coreboot.org/c/flashrom/+/89278
Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>