sb600spi: Check if SPI BAR register has valid value
On systems where AMD ROM Armor is enabled, the SPI BAR register in
LPC PCI configuration space reads as all FFs. Check if the register
reads as all FFs and bail out early if we detect that the SPI base
address is not valid.

TEST=Flashrom does not attempt to access incorrect physical address
on Gigabyte MZ33-AR1 running vendor BIOS with ROM Armor enabled.

Change-Id: I139ea849d6147ea8e8c26ace03627b30f5297267
Signed-off-by: Michał Żygowski <michal.zygowski@3mdeb.com>
Reviewed-on: https://review.coreboot.org/c/flashrom/+/89445
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Anastasia Klimchuk <aklm@chromium.org>