serial.h: Extract serial declarations to a separate header
This patch moves all the declarations relevant to serial into their own
header in include/serial.h.
The corresponding functions implementations are already in serial.c,
so the declarations naturally can be in serial.h
Currently, most of the declarations reside in flash.h making it difficult
to really understand file dependency.

Change-Id: Id0b4d188e2a94dbc4d90747c05eabca39c6b3f26
Signed-off-by: Antonio Vázquez <antoniovazquezblanco@gmail.com>
Reviewed-on: https://review.coreboot.org/c/flashrom/+/89650
Reviewed-by: Anastasia Klimchuk <aklm@chromium.org>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>