dmi.h: Extract DMI declarations to a separate header
This patch moves all the declarations relevant to dmi into their own
header in include/dmi.h.
The corresponding functions implementations are already in dmi.c,
so the declarations naturally can be in dmi.h
Currently, most of the declarations reside in programmer.h making it
difficult to really understand file dependency.

Change-Id: I5fcb14da0b7f478a4f6badc8929c8389a5fa3e3c
Signed-off-by: Antonio Vázquez <antoniovazquezblanco@gmail.com>
Reviewed-on: https://review.coreboot.org/c/flashrom/+/90632
Reviewed-by: Anastasia Klimchuk <aklm@chromium.org>
Reviewed-by: Peter Marheine <pmarheine@chromium.org>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>